employee_records: Also move common.py a folder down
