+++ itou/employee_record/management/commands/transfer_employee_records_updates.py
@@ -12,7 +12,7 @@ from itou.employee_record.models import EmployeeRecordBatch, EmployeeRecordUpdat
 from itou.employee_record.serializers import EmployeeRecordUpdateNotificationBatchSerializer
 from itou.utils.iterators import chunks
 
-from .common import EmployeeRecordTransferCommand
+from ...common_management import EmployeeRecordTransferCommand
 
 
 class Command(EmployeeRecordTransferCommand):
